Move config.json to root directory
Move the config.json file to the root project directory
