+++ Dockerfile
@@ -4,6 +4,7 @@ WORKDIR /app
 
 ADD package.json /app
 ADD package-lock.json /app
+ADD config.json /app
 ADD bot /app/bot
 
 RUN npm install
