+++ README.md
@@ -11,14 +11,14 @@ Or keep reading for self hosting options
 
 ## With Docker
 1. Clone the repo: `git clone https://github.com/PlatypodeCode/CinnamonRoll.git`
-2. Create `config.json` in the `bot` directory (see `sample-config.json` for the layout), and set bot token
+2. Create `config.json` in the same directory as `sample-config.json` (see for the layout), and set bot token: `cp sample-config.json config.json`
 4. Build the bot: `docker-compose build`
 5. Run it: `docker-compose up -d`
 
 ## Manually
 1. Clone the repo: `git clone https://github.com/PlatypodeCode/CinnamonRoll.git`
 2. Run `npm install` to install dependencies
-3. Create `config.json` in the `bot` directory (see `sample-config.json` for the layout), and set bot token
+3. Create `config.json` in the same directory as `sample-config.json` (see for the layout), and set bot token: `cp sample-config.json config.json`
 4. Run a mongo database on localhost and set the database URI in config.json to `mongo://localhost` (or wherever your mongo instance is running)
 5. Run `node bot/main.js` from the top of the directory to run the bot
