+++ bot/commands/say.js
@@ -1,6 +1,6 @@
 const Discord = require('discord.js');
 const utils = require('../util.js')
-const config = require('../config.json');
+const config = require('../../config.json');
 
 module.exports = {
 	name: 'say',
