+++ bot/schemas.js
@@ -1,6 +1,6 @@
 const mongoose = require('mongoose');
 const Schema = mongoose.Schema
-const config = require('./config.json')
+const config = require('../config.json')
 
 
 module.exports = {
