+++ sample-config.json
@@ -1,5 +1,5 @@
 {
-  "defaultPrefix": "!!",
+  "defaultPrefix": "?",
   "token": "aisojwiooajf-dokxjmawiojciokwj-aiodnxoiawnuonfcwuoas",
   "db": "mongodb://db/CinnamonRoll",
   "idCharacters": "abcdefghijklmnopqrstuvwxyz",
